Move `FeedViewController` class to production
